follow up refactors in Rust code
