+++ bindings/java/src/operator.rs
@@ -15,61 +15,34 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use std::str::FromStr;
-use std::time::Duration;
-
 use jni::objects::JByteArray;
 use jni::objects::JClass;
 use jni::objects::JObject;
 use jni::objects::JString;
-use jni::objects::JValue;
-use jni::objects::JValueOwned;
+use jni::sys::jbyteArray;
 use jni::sys::jlong;
 use jni::sys::jobject;
+use jni::sys::jobjectArray;
 use jni::sys::jsize;
 use jni::JNIEnv;
-use opendal::layers::BlockingLayer;
-use opendal::raw::PresignedRequest;
-use opendal::Operator;
-use opendal::Scheme;
 
-use crate::convert::jmap_to_hashmap;
+use opendal::BlockingOperator;
+
 use crate::convert::jstring_to_string;
-use crate::executor::{executor_or_default, get_current_env, Executor};
 use crate::make_entry;
 use crate::make_metadata;
-use crate::make_operator_info;
-use crate::make_presigned_request;
 use crate::Result;
 
+/// # Safety
+///
+/// This function should not be called before the Operator is ready.
 #[no_mangle]
-pub extern "system" fn Java_org_apache_opendal_Operator_constructor(
-    mut env: JNIEnv,
-    _: JClass,
-    executor: *const Executor,
-    scheme: JString,
-    map: JObject,
-) -> jlong {
-    intern_constructor(&mut env, executor, scheme, map).unwrap_or_else(|e| {
-        e.throw(&mut env);
-        0
-    })
-}
-
-fn intern_constructor(
-    env: &mut JNIEnv,
-    executor: *const Executor,
-    scheme: JString,
-    map: JObject,
-) -> Result<jlong> {
-    let scheme = Scheme::from_str(jstring_to_string(env, &scheme)?.as_str())?;
-    let map = jmap_to_hashmap(env, &map)?;
-    let mut op = Operator::via_map(scheme, map)?;
-    if !op.info().full_capability().blocking {
-        let layer = executor_or_default(env, executor)?.enter_with(BlockingLayer::create)?;
-        op = op.layer(layer);
-    }
-    Ok(Box::into_raw(Box::new(op)) as jlong)
+pub unsafe extern "system" fn Java_org_apache_opendal_Operator_disposeInternal(
+    _: JNIEnv,
+    _: JObject,
+    op: *mut BlockingOperator,
+) {
+    drop(Box::from_raw(op));
 }
 
 /// # Safety
@@ -78,8 +51,8 @@ fn intern_constructor(
 #[no_mangle]
 pub unsafe extern "system" fn Java_org_apache_opendal_Operator_duplicate(
     _: JNIEnv,
-    _: JClass,
-    op: *mut Operator,
+    _: JObject,
+    op: *mut BlockingOperator,
 ) -> jlong {
     let op = &mut *op;
     Box::into_raw(Box::new(op.clone())) as jlong
@@ -89,12 +62,23 @@ pub unsafe extern "system" fn Java_org_apache_opendal_Operator_duplicate(
 ///
 /// This function should not be called before the Operator is ready.
 #[no_mangle]
-pub unsafe extern "system" fn Java_org_apache_opendal_Operator_disposeInternal(
-    _: JNIEnv,
-    _: JObject,
-    op: *mut Operator,
-) {
-    drop(Box::from_raw(op));
+pub unsafe extern "system" fn Java_org_apache_opendal_Operator_read(
+    mut env: JNIEnv,
+    _: JClass,
+    op: *mut BlockingOperator,
+    path: JString,
+) -> jbyteArray {
+    intern_read(&mut env, &mut *op, path).unwrap_or_else(|e| {
+        e.throw(&mut env);
+        JByteArray::default().into_raw()
+    })
+}
+
+fn intern_read(env: &mut JNIEnv, op: &mut BlockingOperator, path: JString) -> Result<jbyteArray> {
+    let path = jstring_to_string(env, &path)?;
+    let content = op.read(&path)?.to_bytes();
+    let result = env.byte_array_from_slice(&content)?;
+    Ok(result.into_raw())
 }
 
 /// # Safety
@@ -104,83 +88,24 @@ pub unsafe extern "system" fn Java_org_apache_opendal_Operator_disposeInternal(
 pub unsafe extern "system" fn Java_org_apache_opendal_Operator_write(
     mut env: JNIEnv,
     _: JClass,
-    op: *mut Operator,
-    executor: *const Executor,
+    op: *mut BlockingOperator,
     path: JString,
     content: JByteArray,
-) -> jlong {
-    intern_write(&mut env, op, executor, path, content).unwrap_or_else(|e| {
+) {
+    intern_write(&mut env, &mut *op, path, content).unwrap_or_else(|e| {
         e.throw(&mut env);
-        0
     })
 }
 
 fn intern_write(
     env: &mut JNIEnv,
-    op: *mut Operator,
-    executor: *const Executor,
+    op: &mut BlockingOperator,
     path: JString,
     content: JByteArray,
-) -> Result<jlong> {
-    let op = unsafe { &mut *op };
-    let id = request_id(env)?;
-
+) -> Result<()> {
     let path = jstring_to_string(env, &path)?;
     let content = env.convert_byte_array(content)?;
-
-    executor_or_default(env, executor)?.spawn(async move {
-        let result = do_write(op, path, content).await;
-        complete_future(id, result.map(|_| JValueOwned::Void))
-    });
-
-    Ok(id)
-}
-
-async fn do_write(op: &mut Operator, path: String, content: Vec<u8>) -> Result<()> {
-    Ok(op.write(&path, content).await?)
-}
-
-/// # Safety
-///
-/// This function should not be called before the Operator is ready.
-#[no_mangle]
-pub unsafe extern "system" fn Java_org_apache_opendal_Operator_append(
-    mut env: JNIEnv,
-    _: JClass,
-    op: *mut Operator,
-    executor: *const Executor,
-    path: JString,
-    content: JByteArray,
-) -> jlong {
-    intern_append(&mut env, op, executor, path, content).unwrap_or_else(|e| {
-        e.throw(&mut env);
-        0
-    })
-}
-
-fn intern_append(
-    env: &mut JNIEnv,
-    op: *mut Operator,
-    executor: *const Executor,
-    path: JString,
-    content: JByteArray,
-) -> Result<jlong> {
-    let op = unsafe { &mut *op };
-    let id = request_id(env)?;
-
-    let path = jstring_to_string(env, &path)?;
-    let content = env.convert_byte_array(content)?;
-
-    executor_or_default(env, executor)?.spawn(async move {
-        let result = do_append(op, path, content).await;
-        complete_future(id, result.map(|_| JValueOwned::Void))
-    });
-
-    Ok(id)
-}
-
-async fn do_append(op: &mut Operator, path: String, content: Vec<u8>) -> Result<()> {
-    Ok(op.write_with(&path, content).append(true).await?)
+    Ok(op.write(&path, content)?)
 }
 
 /// # Safety
@@ -190,83 +115,19 @@ async fn do_append(op: &mut Operator, path: String, content: Vec<u8>) -> Result<
 pub unsafe extern "system" fn Java_org_apache_opendal_Operator_stat(
     mut env: JNIEnv,
     _: JClass,
-    op: *mut Operator,
-    executor: *const Executor,
+    op: *mut BlockingOperator,
     path: JString,
-) -> jlong {
-    intern_stat(&mut env, op, executor, path).unwrap_or_else(|e| {
+) -> jobject {
+    intern_stat(&mut env, &mut *op, path).unwrap_or_else(|e| {
         e.throw(&mut env);
-        0
+        JObject::default().into_raw()
     })
 }
 
-fn intern_stat(
-    env: &mut JNIEnv,
-    op: *mut Operator,
-    executor: *const Executor,
-    path: JString,
-) -> Result<jlong> {
-    let op = unsafe { &mut *op };
-    let id = request_id(env)?;
-
+fn intern_stat(env: &mut JNIEnv, op: &mut BlockingOperator, path: JString) -> Result<jobject> {
     let path = jstring_to_string(env, &path)?;
-
-    executor_or_default(env, executor)?.spawn(async move {
-        let result = do_stat(op, path).await;
-        complete_future(id, result.map(JValueOwned::Object))
-    });
-
-    Ok(id)
-}
-
-async fn do_stat<'local>(op: &mut Operator, path: String) -> Result<JObject<'local>> {
-    let metadata = op.stat(&path).await?;
-    let mut env = unsafe { get_current_env() };
-    make_metadata(&mut env, metadata)
-}
-
-/// # Safety
-///
-/// This function should not be called before the Operator is ready.
-#[no_mangle]
-pub unsafe extern "system" fn Java_org_apache_opendal_Operator_read(
-    mut env: JNIEnv,
-    _: JClass,
-    op: *mut Operator,
-    executor: *const Executor,
-    path: JString,
-) -> jlong {
-    intern_read(&mut env, op, executor, path).unwrap_or_else(|e| {
-        e.throw(&mut env);
-        0
-    })
-}
-
-fn intern_read(
-    env: &mut JNIEnv,
-    op: *mut Operator,
-    executor: *const Executor,
-    path: JString,
-) -> Result<jlong> {
-    let op = unsafe { &mut *op };
-    let id = request_id(env)?;
-
-    let path = jstring_to_string(env, &path)?;
-
-    executor_or_default(env, executor)?.spawn(async move {
-        let result = do_read(op, path).await;
-        complete_future(id, result.map(JValueOwned::Object))
-    });
-
-    Ok(id)
-}
-
-async fn do_read<'local>(op: &mut Operator, path: String) -> Result<JObject<'local>> {
-    let content = op.read(&path).await?.to_bytes();
-
-    let env = unsafe { get_current_env() };
-    let result = env.byte_array_from_slice(&content)?;
-    Ok(result.into())
+    let metadata = op.stat(&path)?;
+    Ok(make_metadata(env, metadata)?.into_raw())
 }
 
 /// # Safety
@@ -276,70 +137,17 @@ async fn do_read<'local>(op: &mut Operator, path: String) -> Result<JObject<'loc
 pub unsafe extern "system" fn Java_org_apache_opendal_Operator_delete(
     mut env: JNIEnv,
     _: JClass,
-    op: *mut Operator,
-    executor: *const Executor,
+    op: *mut BlockingOperator,
     path: JString,
-) -> jlong {
-    intern_delete(&mut env, op, executor, path).unwrap_or_else(|e| {
+) {
+    intern_delete(&mut env, &mut *op, path).unwrap_or_else(|e| {
         e.throw(&mut env);
-        0
     })
 }
 
-fn intern_delete(
-    env: &mut JNIEnv,
-    op: *mut Operator,
-    executor: *const Executor,
-    path: JString,
-) -> Result<jlong> {
-    let op = unsafe { &mut *op };
-    let id = request_id(env)?;
-
+fn intern_delete(env: &mut JNIEnv, op: &mut BlockingOperator, path: JString) -> Result<()> {
     let path = jstring_to_string(env, &path)?;
-
-    executor_or_default(env, executor)?.spawn(async move {
-        let result = do_delete(op, path).await;
-        complete_future(id, result.map(|_| JValueOwned::Void))
-    });
-
-    Ok(id)
-}
-
-async fn do_delete(op: &mut Operator, path: String) -> Result<()> {
-    Ok(op.delete(&path).await?)
-}
-
-/// # Safety
-///
-/// This function should not be called before the Operator is ready.
-#[no_mangle]
-pub unsafe extern "system" fn Java_org_apache_opendal_Operator_makeBlockingOp(
-    _: JNIEnv,
-    _: JClass,
-    op: *mut Operator,
-) -> jlong {
-    let op = unsafe { &mut *op };
-    Box::into_raw(Box::new(op.blocking())) as jlong
-}
-
-/// # Safety
-///
-/// This function should not be called before the Operator is ready.
-#[no_mangle]
-pub unsafe extern "system" fn Java_org_apache_opendal_Operator_makeOperatorInfo(
-    mut env: JNIEnv,
-    _: JClass,
-    op: *mut Operator,
-) -> jobject {
-    intern_make_operator_info(&mut env, op).unwrap_or_else(|e| {
-        e.throw(&mut env);
-        JObject::default().into_raw()
-    })
-}
-
-fn intern_make_operator_info(env: &mut JNIEnv, op: *mut Operator) -> Result<jobject> {
-    let op = unsafe { &mut *op };
-    Ok(make_operator_info(env, op.info())?.into_raw())
+    Ok(op.delete(&path)?)
 }
 
 /// # Safety
@@ -349,37 +157,17 @@ fn intern_make_operator_info(env: &mut JNIEnv, op: *mut Operator) -> Result<jobj
 pub unsafe extern "system" fn Java_org_apache_opendal_Operator_createDir(
     mut env: JNIEnv,
     _: JClass,
-    op: *mut Operator,
-    executor: *const Executor,
+    op: *mut BlockingOperator,
     path: JString,
-) -> jlong {
-    intern_create_dir(&mut env, op, executor, path).unwrap_or_else(|e| {
+) {
+    intern_create_dir(&mut env, &mut *op, path).unwrap_or_else(|e| {
         e.throw(&mut env);
-        0
     })
 }
 
-fn intern_create_dir(
-    env: &mut JNIEnv,
-    op: *mut Operator,
-    executor: *const Executor,
-    path: JString,
-) -> Result<jlong> {
-    let op = unsafe { &mut *op };
-    let id = request_id(env)?;
-
+fn intern_create_dir(env: &mut JNIEnv, op: &mut BlockingOperator, path: JString) -> Result<()> {
     let path = jstring_to_string(env, &path)?;
-
-    executor_or_default(env, executor)?.spawn(async move {
-        let result = do_create_dir(op, path).await;
-        complete_future(id, result.map(|_| JValueOwned::Void))
-    });
-
-    Ok(id)
-}
-
-async fn do_create_dir(op: &mut Operator, path: String) -> Result<()> {
-    Ok(op.create_dir(&path).await?)
+    Ok(op.create_dir(&path)?)
 }
 
 /// # Safety
@@ -389,40 +177,25 @@ async fn do_create_dir(op: &mut Operator, path: String) -> Result<()> {
 pub unsafe extern "system" fn Java_org_apache_opendal_Operator_copy(
     mut env: JNIEnv,
     _: JClass,
-    op: *mut Operator,
-    executor: *const Executor,
+    op: *mut BlockingOperator,
     source_path: JString,
     target_path: JString,
-) -> jlong {
-    intern_copy(&mut env, op, executor, source_path, target_path).unwrap_or_else(|e| {
+) {
+    intern_copy(&mut env, &mut *op, source_path, target_path).unwrap_or_else(|e| {
         e.throw(&mut env);
-        0
     })
 }
 
 fn intern_copy(
     env: &mut JNIEnv,
-    op: *mut Operator,
-    executor: *const Executor,
+    op: &mut BlockingOperator,
     source_path: JString,
     target_path: JString,
-) -> Result<jlong> {
-    let op = unsafe { &mut *op };
-    let id = request_id(env)?;
-
+) -> Result<()> {
     let source_path = jstring_to_string(env, &source_path)?;
     let target_path = jstring_to_string(env, &target_path)?;
 
-    executor_or_default(env, executor)?.spawn(async move {
-        let result = do_copy(op, source_path, target_path).await;
-        complete_future(id, result.map(|_| JValueOwned::Void))
-    });
-
-    Ok(id)
-}
-
-async fn do_copy(op: &mut Operator, source_path: String, target_path: String) -> Result<()> {
-    Ok(op.copy(&source_path, &target_path).await?)
+    Ok(op.copy(&source_path, &target_path)?)
 }
 
 /// # Safety
@@ -432,40 +205,25 @@ async fn do_copy(op: &mut Operator, source_path: String, target_path: String) ->
 pub unsafe extern "system" fn Java_org_apache_opendal_Operator_rename(
     mut env: JNIEnv,
     _: JClass,
-    op: *mut Operator,
-    executor: *const Executor,
+    op: *mut BlockingOperator,
     source_path: JString,
     target_path: JString,
-) -> jlong {
-    intern_rename(&mut env, op, executor, source_path, target_path).unwrap_or_else(|e| {
+) {
+    intern_rename(&mut env, &mut *op, source_path, target_path).unwrap_or_else(|e| {
         e.throw(&mut env);
-        0
     })
 }
 
 fn intern_rename(
     env: &mut JNIEnv,
-    op: *mut Operator,
-    executor: *const Executor,
+    op: &mut BlockingOperator,
     source_path: JString,
     target_path: JString,
-) -> Result<jlong> {
-    let op = unsafe { &mut *op };
-    let id = request_id(env)?;
-
+) -> Result<()> {
     let source_path = jstring_to_string(env, &source_path)?;
     let target_path = jstring_to_string(env, &target_path)?;
 
-    executor_or_default(env, executor)?.spawn(async move {
-        let result = do_rename(op, source_path, target_path).await;
-        complete_future(id, result.map(|_| JValueOwned::Void))
-    });
-
-    Ok(id)
-}
-
-async fn do_rename(op: &mut Operator, source_path: String, target_path: String) -> Result<()> {
-    Ok(op.rename(&source_path, &target_path).await?)
+    Ok(op.rename(&source_path, &target_path)?)
 }
 
 /// # Safety
@@ -475,37 +233,18 @@ async fn do_rename(op: &mut Operator, source_path: String, target_path: String) 
 pub unsafe extern "system" fn Java_org_apache_opendal_Operator_removeAll(
     mut env: JNIEnv,
     _: JClass,
-    op: *mut Operator,
-    executor: *const Executor,
+    op: *mut BlockingOperator,
     path: JString,
-) -> jlong {
-    intern_remove_all(&mut env, op, executor, path).unwrap_or_else(|e| {
+) {
+    intern_remove_all(&mut env, &mut *op, path).unwrap_or_else(|e| {
         e.throw(&mut env);
-        0
     })
 }
 
-fn intern_remove_all(
-    env: &mut JNIEnv,
-    op: *mut Operator,
-    executor: *const Executor,
-    path: JString,
-) -> Result<jlong> {
-    let op = unsafe { &mut *op };
-    let id = request_id(env)?;
-
+fn intern_remove_all(env: &mut JNIEnv, op: &mut BlockingOperator, path: JString) -> Result<()> {
     let path = jstring_to_string(env, &path)?;
 
-    executor_or_default(env, executor)?.spawn(async move {
-        let result = do_remove_all(op, path).await;
-        complete_future(id, result.map(|_| JValueOwned::Void))
-    });
-
-    Ok(id)
-}
-
-async fn do_remove_all(op: &mut Operator, path: String) -> Result<()> {
-    Ok(op.remove_all(&path).await?)
+    Ok(op.remove_all(&path)?)
 }
 
 /// # Safety
@@ -515,39 +254,19 @@ async fn do_remove_all(op: &mut Operator, path: String) -> Result<()> {
 pub unsafe extern "system" fn Java_org_apache_opendal_Operator_list(
     mut env: JNIEnv,
     _: JClass,
-    op: *mut Operator,
-    executor: *const Executor,
+    op: *mut BlockingOperator,
     path: JString,
-) -> jlong {
-    intern_list(&mut env, op, executor, path).unwrap_or_else(|e| {
+) -> jobjectArray {
+    intern_list(&mut env, &mut *op, path).unwrap_or_else(|e| {
         e.throw(&mut env);
-        0
+        JObject::default().into_raw()
     })
 }
 
-fn intern_list(
-    env: &mut JNIEnv,
-    op: *mut Operator,
-    executor: *const Executor,
-    path: JString,
-) -> Result<jlong> {
-    let op = unsafe { &mut *op };
-    let id = request_id(env)?;
-
+fn intern_list(env: &mut JNIEnv, op: &mut BlockingOperator, path: JString) -> Result<jobjectArray> {
     let path = jstring_to_string(env, &path)?;
+    let obs = op.list(&path)?;
 
-    executor_or_default(env, executor)?.spawn(async move {
-        let result = do_list(op, path).await;
-        complete_future(id, result.map(JValueOwned::Object))
-    });
-
-    Ok(id)
-}
-
-async fn do_list<'local>(op: &mut Operator, path: String) -> Result<JObject<'local>> {
-    let obs = op.list(&path).await?;
-
-    let mut env = unsafe { get_current_env() };
     let jarray = env.new_object_array(
         obs.len() as jsize,
         "org/apache/opendal/Entry",
@@ -555,227 +274,9 @@ async fn do_list<'local>(op: &mut Operator, path: String) -> Result<JObject<'loc
     )?;
 
     for (idx, entry) in obs.iter().enumerate() {
-        let entry = make_entry(&mut env, entry.to_owned())?;
+        let entry = make_entry(env, entry.to_owned())?;
         env.set_object_array_element(&jarray, idx as jsize, entry)?;
     }
 
-    Ok(jarray.into())
-}
-
-/// # Safety
-///
-/// This function should not be called before the Operator is ready.
-#[no_mangle]
-pub unsafe extern "system" fn Java_org_apache_opendal_Operator_presignRead(
-    mut env: JNIEnv,
-    _: JClass,
-    op: *mut Operator,
-    executor: *const Executor,
-    path: JString,
-    expire: jlong,
-) -> jlong {
-    intern_presign_read(&mut env, op, executor, path, expire).unwrap_or_else(|e| {
-        e.throw(&mut env);
-        0
-    })
-}
-
-fn intern_presign_read(
-    env: &mut JNIEnv,
-    op: *mut Operator,
-    executor: *const Executor,
-    path: JString,
-    expire: jlong,
-) -> Result<jlong> {
-    let op = unsafe { &mut *op };
-    let id = request_id(env)?;
-
-    let path = jstring_to_string(env, &path)?;
-    let expire = Duration::from_nanos(expire as u64);
-
-    executor_or_default(env, executor)?.spawn(async move {
-        let result = do_presign_read(op, path, expire).await;
-        let mut env = unsafe { get_current_env() };
-        let result = result.and_then(|req| make_presigned_request(&mut env, req));
-        complete_future(id, result.map(JValueOwned::Object))
-    });
-
-    Ok(id)
-}
-
-async fn do_presign_read(
-    op: &mut Operator,
-    path: String,
-    expire: Duration,
-) -> Result<PresignedRequest> {
-    Ok(op.presign_read(&path, expire).await?)
-}
-
-/// # Safety
-///
-/// This function should not be called before the Operator is ready.
-#[no_mangle]
-pub unsafe extern "system" fn Java_org_apache_opendal_Operator_presignWrite(
-    mut env: JNIEnv,
-    _: JClass,
-    op: *mut Operator,
-    executor: *const Executor,
-    path: JString,
-    expire: jlong,
-) -> jlong {
-    intern_presign_write(&mut env, op, executor, path, expire).unwrap_or_else(|e| {
-        e.throw(&mut env);
-        0
-    })
-}
-
-fn intern_presign_write(
-    env: &mut JNIEnv,
-    op: *mut Operator,
-    executor: *const Executor,
-    path: JString,
-    expire: jlong,
-) -> Result<jlong> {
-    let op = unsafe { &mut *op };
-    let id = request_id(env)?;
-
-    let path = jstring_to_string(env, &path)?;
-    let expire = Duration::from_nanos(expire as u64);
-
-    executor_or_default(env, executor)?.spawn(async move {
-        let result = do_presign_write(op, path, expire).await;
-        let mut env = unsafe { get_current_env() };
-        let result = result.and_then(|req| make_presigned_request(&mut env, req));
-        complete_future(id, result.map(JValueOwned::Object))
-    });
-
-    Ok(id)
-}
-
-async fn do_presign_write(
-    op: &mut Operator,
-    path: String,
-    expire: Duration,
-) -> Result<PresignedRequest> {
-    Ok(op.presign_write(&path, expire).await?)
-}
-
-/// # Safety
-///
-/// This function should not be called before the Operator is ready.
-#[no_mangle]
-pub unsafe extern "system" fn Java_org_apache_opendal_Operator_presignStat(
-    mut env: JNIEnv,
-    _: JClass,
-    op: *mut Operator,
-    executor: *const Executor,
-    path: JString,
-    expire: jlong,
-) -> jlong {
-    intern_presign_stat(&mut env, op, executor, path, expire).unwrap_or_else(|e| {
-        e.throw(&mut env);
-        0
-    })
-}
-
-fn intern_presign_stat(
-    env: &mut JNIEnv,
-    op: *mut Operator,
-    executor: *const Executor,
-    path: JString,
-    expire: jlong,
-) -> Result<jlong> {
-    let op = unsafe { &mut *op };
-    let id = request_id(env)?;
-
-    let path = jstring_to_string(env, &path)?;
-    let expire = Duration::from_nanos(expire as u64);
-
-    executor_or_default(env, executor)?.spawn(async move {
-        let result = do_presign_stat(op, path, expire).await;
-        let mut env = unsafe { get_current_env() };
-        let result = result.and_then(|req| make_presigned_request(&mut env, req));
-        complete_future(id, result.map(JValueOwned::Object))
-    });
-
-    Ok(id)
-}
-
-async fn do_presign_stat(
-    op: &mut Operator,
-    path: String,
-    expire: Duration,
-) -> Result<PresignedRequest> {
-    Ok(op.presign_stat(&path, expire).await?)
-}
-
-fn make_object<'local>(
-    env: &mut JNIEnv<'local>,
-    value: JValueOwned<'local>,
-) -> Result<JObject<'local>> {
-    let o = match value {
-        JValueOwned::Object(o) => o,
-        JValueOwned::Byte(_) => env.new_object("java/lang/Long", "(B)V", &[value.borrow()])?,
-        JValueOwned::Char(_) => env.new_object("java/lang/Char", "(C)V", &[value.borrow()])?,
-        JValueOwned::Short(_) => env.new_object("java/lang/Short", "(S)V", &[value.borrow()])?,
-        JValueOwned::Int(_) => env.new_object("java/lang/Integer", "(I)V", &[value.borrow()])?,
-        JValueOwned::Long(_) => env.new_object("java/lang/Long", "(J)V", &[value.borrow()])?,
-        JValueOwned::Bool(_) => env.new_object("java/lang/Boolean", "(Z)V", &[value.borrow()])?,
-        JValueOwned::Float(_) => env.new_object("java/lang/Float", "(F)V", &[value.borrow()])?,
-        JValueOwned::Double(_) => env.new_object("java/lang/Double", "(D)V", &[value.borrow()])?,
-        JValueOwned::Void => JObject::null(),
-    };
-    Ok(o)
-}
-
-fn complete_future(id: jlong, result: Result<JValueOwned>) {
-    try_complete_future(id, result).expect("complete future must succeed");
-}
-
-fn try_complete_future(id: jlong, result: Result<JValueOwned>) -> Result<()> {
-    let mut env = unsafe { get_current_env() };
-    let future = get_future(&mut env, id)?;
-    match result {
-        Ok(result) => {
-            let result = make_object(&mut env, result)?;
-            env.call_method(
-                future,
-                "complete",
-                "(Ljava/lang/Object;)Z",
-                &[JValue::Object(&result)],
-            )?
-        }
-        Err(err) => {
-            let exception = err.to_exception(&mut env)?;
-            env.call_method(
-                future,
-                "completeExceptionally",
-                "(Ljava/lang/Throwable;)Z",
-                &[JValue::Object(&exception)],
-            )?
-        }
-    };
-    Ok(())
-}
-
-fn request_id(env: &mut JNIEnv) -> Result<jlong> {
-    Ok(env
-        .call_static_method(
-            "org/apache/opendal/Operator$AsyncRegistry",
-            "requestId",
-            "()J",
-            &[],
-        )?
-        .j()?)
-}
-
-fn get_future<'local>(env: &mut JNIEnv<'local>, id: jlong) -> Result<JObject<'local>> {
-    Ok(env
-        .call_static_method(
-            "org/apache/opendal/Operator$AsyncRegistry",
-            "get",
-            "(J)Ljava/util/concurrent/CompletableFuture;",
-            &[JValue::Long(id)],
-        )?
-        .l()?)
+    Ok(jarray.into_raw())
 }
